# speed sam.ParseCigar