```
benchmark                 old ns/op     new ns/op     delta
BenchmarkParseCigar-4     18076         3956          -78.11%
```
I'm using sam.ParseCigar to parse the cigar in SA tags in bams from oxford nanopore and the cigar strings are very long. This improves the performance of the full program somewhat. 
It's specialized, but it is safe since there are already checks the '0' <= c <= '9' for all characters we're converting to int.