@brentp 2016-11-14
I made those changes.